fix gung-detail 컴포넌트 분리
